[IMP] mail_post_defer: defer more mails
In the previous version of this module, only emails sent through the chatter were explicitly deferred. However, there are more mails that are auto-generated by Odoo and could benefit a lot from the deferring system. One of such cases is the automatic assignation mail.

To cover that and more cases, now the auto-deferring is done at a lower level. If we get some hint that the mail sending is forced, or that it's already deferred, we skip the machinery. If we get nothing about all that, then we just add our defaults to avoid force-sending and defer by default.

The result is that more Odoo internal systems will use the mail queue, and thus there'll be less blocking.

As an exception, when running `_send_notifications()`, we force-disable the deferring system. That is executed by a cron, or by the user when force-sending a notification. So, only in that case the queue should be skipped by default.

@moduon MT-6204